Add multiple endpoint support
Add an Endpoint class as a part of the resource module. An Endpoint can be
optionally passed to any of of the resource objects to allow them to
access a different DCM endpoint then the one specified either by
environment variables or in configuration files as loaded by the config
class.

- Issue #210 heavily inspired this feature, thanks @thijs-creemers
- Fixes Issue #207