Fix runtime creation of property access methods
In the event that we are missing an access method for a property that is
present in the json returned from the DCM API, the access methods are
created using Resource.add_property. This was implemented previously
but was broken. This commit also adds a nosetest test_user_arbitrary
which inserts unexpected properties in a mock API response. Fixes #229