Handle Endpoints in classmethods and Job waiting
There are a few places where the Job class is used to wait for
completion of operations inside other classes. In these cases we have to
make sure that we use the same endpoint as the enclosing object
instance. There were also a few classmethods particularly associated
with volumes that were missed in the initial conversion to endpoints.
Note that the default behavior is still preserved.